requirements.txt
#LangChain Core
langchain
langchain-core

#OpenAI Integration
langchain-openai
openai

#Anthropic Integration
langchain-anthropic

#Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

#Hugging Face Integration 
langchain-huggingface
transformers
huggingface-hub

#Environment Variable Management
python-dotenv

#Machine Learning Utilities
numpy
scikit-learn

jinja
torch
sentence-transformers
streamlit
pydantic
sentencepiece
grandalf
langchain_community
pypdf
beautifulsoup4
langchain_experimental
tiktoken
chromadb
tqdm
hf_xet
langchain_chroma
faiss-cpu
wikipedia
youtube-transcript-api